* 处理警告

+++ Vurig Calendar/NSDate+convenience.h
@@ -9,14 +9,14 @@
 
 @interface NSDate (Convenience)
 
--(NSDate *)offsetMonth:(int)numMonths;
--(NSDate *)offsetDay:(int)numDays;
--(NSDate *)offsetHours:(int)hours;
--(int)numDaysInMonth;
--(int)firstWeekDayInMonth;
--(int)year;
--(int)month;
--(int)day;
+-(NSDate *)offsetMonth:(NSInteger)numMonths;
+-(NSDate *)offsetDay:(NSInteger)numDays;
+-(NSDate *)offsetHours:(NSInteger)hours;
+-(NSInteger)numDaysInMonth;
+-(NSInteger)firstWeekDayInMonth;
+-(NSInteger)year;
+-(NSInteger)month;
+-(NSInteger)day;
 
 +(NSDate *)dateStartOfDay:(NSDate *)date;
 +(NSDate *)dateStartOfWeek;
