+++ Vurig Calendar/NSDate+convenience.m
@@ -9,7 +9,7 @@
 
 @implementation NSDate (Convenience)
 
--(int)year {
+- (NSInteger)year {
     NSCalendar *gregorian = [[NSCalendar alloc]
                              initWithCalendarIdentifier:NSGregorianCalendar];
     NSDateComponents *components = [gregorian components:NSYearCalendarUnit fromDate:self];
@@ -17,21 +17,21 @@
 }
 
 
--(int)month {
+- (NSInteger)month {
     NSCalendar *gregorian = [[NSCalendar alloc]
                              initWithCalendarIdentifier:NSGregorianCalendar];
     NSDateComponents *components = [gregorian components:NSMonthCalendarUnit fromDate:self];
     return [components month];
 }
 
--(int)day {
+- (NSInteger)day {
     NSCalendar *gregorian = [[NSCalendar alloc]
                              initWithCalendarIdentifier:NSGregorianCalendar];
     NSDateComponents *components = [gregorian components:NSDayCalendarUnit fromDate:self];
     return [components day];
 }
 
--(int)firstWeekDayInMonth {
+- (NSInteger)firstWeekDayInMonth {
     NSCalendar *gregorian = [[NSCalendar alloc]
                              initWithCalendarIdentifier:NSGregorianCalendar];
     [gregorian setFirstWeekday:2]; //monday is first day
@@ -45,7 +45,7 @@
     return [gregorian ordinalityOfUnit:NSWeekdayCalendarUnit inUnit:NSWeekCalendarUnit forDate:newDate];
 }
 
--(NSDate *)offsetMonth:(int)numMonths {
+- (NSDate *)offsetMonth:(NSInteger)numMonths {
     NSCalendar *gregorian = [[NSCalendar alloc]
                              initWithCalendarIdentifier:NSGregorianCalendar];
     [gregorian setFirstWeekday:2]; //monday is first day
@@ -58,7 +58,7 @@
                                                    toDate:self options:0];
 }
 
--(NSDate *)offsetHours:(int)hours {
+- (NSDate *)offsetHours:(NSInteger)hours {
     NSCalendar *gregorian = [[NSCalendar alloc]
                              initWithCalendarIdentifier:NSGregorianCalendar];
     [gregorian setFirstWeekday:2]; //monday is first day
@@ -71,7 +71,7 @@
                                       toDate:self options:0];
 }
 
--(NSDate *)offsetDay:(int)numDays {
+- (NSDate *)offsetDay:(NSInteger)numDays {
     NSCalendar *gregorian = [[NSCalendar alloc]
                              initWithCalendarIdentifier:NSGregorianCalendar];
     [gregorian setFirstWeekday:2]; //monday is first day
@@ -87,7 +87,7 @@
 
 
 
--(int)numDaysInMonth {
+- (NSInteger)numDaysInMonth {
     NSCalendar *cal = [NSCalendar currentCalendar];
     NSRange rng = [cal rangeOfUnit:NSDayCalendarUnit inUnit:NSMonthCalendarUnit forDate:self];
     NSUInteger numberOfDaysInMonth = rng.length;
